Move WeightInfo, Config, Events into ExternalAgent pallet.
