+++ pallets/common/Cargo.toml
@@ -12,6 +12,7 @@ polymesh-primitives-derive = { workspace = true, default-features = false }
 pallet-base = { workspace = true, default-features = false  }
 pallet-permissions = { workspace = true, default-features = false  }
 pallet-identity = { workspace = true, default-features = false  }
+pallet-external-agents = { workspace = true, default-features = false  }
 
 # Other
 serde = { version = "1.0.112", default-features = false }
@@ -58,6 +59,7 @@ std = [
 	"pallet-base/std",
 	"pallet-permissions/std",
 	"pallet-identity/std",
+	"pallet-external-agents/std",
 	"polymesh-primitives/std",
 	"serde/std",
 	"serde_derive",
@@ -72,5 +74,6 @@ runtime-benchmarks = [
 	"frame-benchmarking",
 	"pallet-permissions/runtime-benchmarks",
 	"pallet-identity/runtime-benchmarks",
+	"pallet-external-agents/runtime-benchmarks",
 	"schnorrkel",
 ]
